pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seafoam-cicada"
version = "0.1.0"
description = "Token-optimized LangGraph ticket triage orchestrator"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

authors = [
    {name = "Development Team"}
]

keywords = ["langgraph", "ticket-triage", "orchestration", "llm"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core API
    "fastapi==0.115.0",
    "uvicorn==0.30.6",
    "pydantic==2.9.2",
    
    # LangGraph & LangChain
    "langgraph==1.0.5",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    
    # Observability
    "langsmith>=0.1.0",
    
    # Environment
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/example/seafoam-cicada"
Documentation = "https://github.com/example/seafoam-cicada#readme"
Repository = "https://github.com/example/seafoam-cicada.git"

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W"]
ignore = ["E501", "W503"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=app --cov-report=html"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/__main__.py"]
